+++ src/platform-element-plus/App.tsx
@@ -8,7 +8,6 @@ import { ElMessage } from "element-plus";
 import { Global, createAtom } from "@vue-start/css";
 import { IRequestActor } from "@vue-start/request";
 import { createCssVar } from "@/style/theme";
-
 import { ChengOpe } from "@/layout/cheng";
 
 const showMsg = (opts: any) => {