+++ src/platform-element-plus/component.tsx
@@ -1,5 +1,5 @@
 import { App } from "@vue/runtime-core";
-import { ElButton, ElInputNumber, ElMessageBox, ElIcon,ElScrollbar, ElTree, ElInput, ElSwitch } from "element-plus";
+import { ElButton, ElInputNumber, ElMessageBox, ElIcon, ElScrollbar, ElTree, ElInput, ElSwitch } from "element-plus";
 import { ArrowLeftBold } from "@element-plus/icons-vue";
 import {
   ProLoading,