+++ @vue-start/pro/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 0.5.52
+
+- fix: ProTable header 有效性 css。通过 table css 取： pro-table.has-header .pro-table-toolbar；通过直接属性取：pro-table-toolbar-valid；
+
+# 0.5.51
+
+- fix: 左侧菜单收起 css 类精确，由.pro-layout.mini 改为 .pro-layout.has-left-menu.mini
+- fix: ProTheme 新增 lightThemeToken、darkThemeToken，支持 ThemeToken 自定义
+- fix: FormItem label 插槽重写 bug 修复；
+- fix: ProLayout routeOpts（ProRouterView 组件） 支持 tagsCache 配置；ProLayout menus 项新增 keep 配置（加入 keep-live 缓存）；
+
+# 0.5.50
+
+- fix: ProCurdModule ref 穿透；
+- fix: ProCurdPage as 默认 div；
+
 # 0.5.49
 
 - refactor: ProTheme 支持自模块使用；新增模式（暗黑）；